Refactor commands to bundled Node.js GitHub Action
