+++ commands/action.yml
@@ -11,5 +11,5 @@ inputs:
     description: Grafana Cloud metrics api key
     required: false
 runs:
-  using: 'node20'
-  main: 'index.js'
+  using: 'node24'
+  main: 'dist/index.js'
