+++ commands/package.json
@@ -1,0 +1,26 @@
+{
+	"name": "grafana-github-actions-commands",
+	"private": true,
+	"version": "1.0.0",
+	"type": "module",
+	"scripts": {
+		"build": "rollup --config rollup.config.js"
+	},
+	"dependencies": {
+		"@actions/github": "^2.1.1",
+		"@octokit/webhooks": "^7.12.2"
+	},
+	"devDependencies": {
+		"@rollup/plugin-commonjs": "^29.0.2",
+		"@rollup/plugin-json": "^6.1.0",
+		"@rollup/plugin-node-resolve": "^16.0.3",
+		"@rollup/plugin-typescript": "^12.3.0",
+		"@types/chai": "^4.2.10",
+		"@types/jest": "^27.4.0",
+		"@types/node": "^20.9.2",
+		"chai": "^4.2.0",
+		"jest": "^27.4",
+		"rollup": "^4.60.1",
+		"typescript": "^6.0.2"
+	}
+}
