+++ tsconfig.json
@@ -11,5 +11,5 @@
 		"lib": ["es2021"],
 		"skipLibCheck": true
 	},
-	"exclude": ["**/*.test.ts", "**/vm-filesystem/**", "__mocks__/**/*", "gha-otel-export/**", "pr-checks/**"]
+	"exclude": ["**/*.test.ts", "**/vm-filesystem/**", "__mocks__/**/*", "gha-otel-export/**", "pr-checks/**", "commands/**"]
 }
